Moves entrypoint to app module
