+++ app/build.gradle.kts
@@ -23,15 +23,21 @@ dependencies {
     implementation(projects.platform.core.coreRest)
     implementation(projects.platform.core.corePersistance)
     implementation(projects.platform.core.coreNavigator)
+    implementation(projects.platform.common.commonAndroid)
     implementation(projects.platform.common.commonKodein)
     implementation(projects.platform.common.commonStatic)
     implementation(projects.features.facts)
-    implementation(projects.features.onboarding)
     implementation(projects.features.search)
+
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.5.2")
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core-jvm:1.5.2")
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-android:1.5.2")
 
     implementation("androidx.core:core:1.6.0")
     implementation("androidx.core:core-ktx:1.6.0")
     implementation("androidx.appcompat:appcompat:1.3.1")
+    implementation("androidx.lifecycle:lifecycle-common:2.3.1")
+    implementation("androidx.lifecycle:lifecycle-runtime-ktx:2.3.1")
     implementation("org.kodein.di:kodein-di-jvm:7.9.0")
     implementation("org.kodein.type:kodein-type-jvm:1.11.0")
     implementation("com.squareup.okhttp3:okhttp:4.9.2")
