+++ app/src/main/java/io/dotanuki/demos/norris/SplashActivity.kt
@@ -1,13 +1,13 @@
-package io.dotanuki.norris.onboarding
+package io.dotanuki.demos.norris
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.lifecycleScope
+import io.dotanuki.demos.norris.databinding.ActivitySplashBinding
 import io.dotanuki.norris.common.android.selfBind
 import io.dotanuki.norris.common.android.viewBinding
 import io.dotanuki.norris.navigator.Navigator
 import io.dotanuki.norris.navigator.Screen
-import io.dotanuki.norris.onboarding.databinding.ActivityOnboardingBinding
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import org.kodein.di.DIAware
@@ -15,11 +15,11 @@ import org.kodein.di.instance
 import android.R.anim.fade_in as FadeIn
 import android.R.anim.fade_out as FadeOut
 
-class OnboardingActivity : AppCompatActivity(), DIAware {
+class SplashActivity : AppCompatActivity(), DIAware {
 
     override val di by selfBind()
 
-    private val viewBindings by viewBinding(ActivityOnboardingBinding::inflate)
+    private val viewBindings by viewBinding(ActivitySplashBinding::inflate)
     private val navigator by instance<Navigator>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
