+++ app/src/main/res/layout/activity_splash.xml
