+++ app/src/main/res/values/strings_splash.xml
