--- features/onboarding/build.gradle.kts
@@ -1,20 +1,0 @@
-
-plugins {
-    id("norris.modules.android.feature")
-}
-
-dependencies {
-    implementation(projects.platform.common.commonAndroid)
-    implementation(projects.platform.common.commonKodein)
-    implementation(projects.platform.common.commonStatic)
-    implementation(projects.platform.core.coreNavigator)
-
-    implementation("org.kodein.di:kodein-di-jvm:7.9.0")
-    implementation("org.kodein.type:kodein-type-jvm:1.11.0")
-    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.5.2")
-    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core-jvm:1.5.2")
-    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-android:1.5.2")
-    implementation("androidx.appcompat:appcompat:1.3.1")
-    implementation("androidx.lifecycle:lifecycle-common:2.3.1")
-    implementation("androidx.lifecycle:lifecycle-runtime-ktx:2.3.1")
-}
