--- features/onboarding/src/main/AndroidManifest.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="io.dotanuki.norris.onboarding">
-
-    <application>
-
-        <activity
-            android:name=".OnboardingActivity"
-            android:exported="true"
-            android:screenOrientation="portrait"/>
-
-    </application>
-
-</manifest>
